feat: modularized avatar components (reusability)
